echo/echo_backend/requirements.txt
django>=4.2.0
channels>=4.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
django-filter>=23.3
drf-yasg>=1.21.7
Pillow>=10.0.0
django-cleanup>=8.0.0
django-storages>=1.14.2
boto3>=1.28.44
psycopg2>=2.9.7
django-rest-passwordreset>=1.3.0
django-rest-knox>=4.2.0 